Reseting gauge on backend metrics
After some value is setted to metricss `postgresql_backends_by_wait_event_type_total`
and `postgresql_backends_by_state_total`, the value is not updated if the query returns
no rows.

To fix that, the gauge is reseted every time when the query is executeted, thus grants
that the old value cleared.